Add files via upload
The code allows me to plot the frequency response function to evaluate the amplitude of input and output signals of a filter amplifier. It also allows me to evaluate the phase shift. Everything is associated to the cutoff frequency with which it works. For this case, the data are from a 1000 Hz HighPass with input signals from 500 to 1500 Hz with 100 Hz jumps. The phase diagram is not obtained correctly.